moved the journal entries into a journal Entry folder then refactured the index form for entries into javascript to appear on dom and be functional, record button still not working
