+++ scripts/JournalEntry/JournalDataProvider.js
