+++ scripts/main.js
@@ -1,2 +1,5 @@
-import { EntryListComponent } from './JournalEntryList.js'
+import {JournalFormComponent} from "./JournalForm/form.js"
+import { EntryListComponent } from './JournalEntry/JournalEntryList.js'
+
+JournalFormComponent()
 EntryListComponent()
